Update mail.md (#86)
* Update mail.md

Change the example connection in the Queue method from `high` to `redis` to make it more obvious that the value should be a valid connection defined in _config/queue.go_.

* Update mail.md

Changing other places where `high` was used as a value in the `Connection()` method.

* Update mail.md

Update the Chinese version of the mail docs to match the English version.